+++ NEWS.md
@@ -1,4 +1,4 @@
-## v0.1.9.9001 
+## v0.2.0
 
 * Method `Duplicates$detectDuplicates()` now has arguments `n` (passed into 
 polmineR::ngrams) and `character_selection`. Values were hard-coded previously.
@@ -9,6 +9,8 @@ Default values are aligned with Kliche et al. 2014.
 now and will be available for `plpr_subcorpus` by inheritance #3.
 * An outdated example für the `Duplicates` class using a `duplicates()` method
 has been removed #7.
+* `Duplicates$detectDuplicates()` works for `n` = 0 (same-day comparisons only)
+#11.
 
 ## v0.1.9